# api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "新猿译码 Gateway API",
    "version": "0.1.0",
    "description": "新猿译码一期 Gateway 的公开只读 API 契约。"
  },
  "servers": [
    {
      "url": "/",
      "description": "同源 Gateway"
    }
  ],
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "summary": "注册并创建会话",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } } }
        },
        "responses": {
          "201": {
            "description": "注册成功",
            "headers": { "Set-Cookie": { "schema": { "type": "string" } } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } }
          },
          "409": { "description": "邮箱已注册", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "422": { "description": "注册字段无效", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "429": { "description": "请求过于频繁", "headers": { "Retry-After": { "schema": { "type": "integer" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "邮箱密码登录",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "headers": { "Set-Cookie": { "schema": { "type": "string" } } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } }
          },
          "401": { "description": "凭据无效", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "429": { "description": "请求过于频繁", "headers": { "Retry-After": { "schema": { "type": "integer" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "summary": "退出并销毁会话",
        "operationId": "logout",
        "security": [{ "cookieAuth": [] }],
        "responses": { "204": { "description": "已退出" } }
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "summary": "退出当前用户的全部设备",
        "operationId": "logoutAll",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "204": { "description": "全部会话已失效" },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/sessions": {
      "get": {
        "summary": "获取当前用户的有效登录会话",
        "operationId": "listAuthSessions",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "有效登录会话列表", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SessionListResponse" } } } },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/sessions/{sessionID}": {
      "delete": {
        "summary": "撤销指定登录会话",
        "operationId": "revokeAuthSession",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "sessionID", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "指定会话已撤销" },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "会话不存在或不属于当前用户", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/password": {
      "put": {
        "summary": "修改邮箱账号密码并撤销其他设备会话",
        "operationId": "changePassword",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PasswordChangeRequest" } } }
        },
        "responses": {
          "204": { "description": "密码已修改，其他会话已撤销" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "description": "未登录或当前密码错误", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "422": { "description": "账号未设置密码或新密码不符合要求", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "429": { "description": "请求过于频繁", "headers": { "Retry-After": { "schema": { "type": "integer" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/password-reset/request": {
      "post": {
        "summary": "请求密码重置邮件",
        "operationId": "requestPasswordReset",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PasswordResetRequest" } } }
        },
        "responses": {
          "202": { "description": "请求已接受；无论邮箱是否存在均返回相同响应", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AcceptedResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "422": { "description": "邮箱格式无效", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "429": { "description": "请求过于频繁", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "503": { "description": "邮件投递尚未配置", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/password-reset/confirm": {
      "post": {
        "summary": "使用一次性令牌设置新密码",
        "operationId": "confirmPasswordReset",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PasswordResetConfirmRequest" } } }
        },
        "responses": {
          "204": { "description": "密码已重置且全部登录会话已撤销" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "422": { "description": "令牌无效、已使用、已过期或新密码无效", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "429": { "description": "请求过于频繁", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "summary": "获取当前用户",
        "operationId": "getCurrentUser",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "当前用户", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } } },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "patch": {
        "summary": "更新当前用户资料",
        "operationId": "updateCurrentUser",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProfileUpdateRequest" } } }
        },
        "responses": {
          "200": { "description": "更新后的当前用户", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "422": { "description": "昵称不符合要求", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/oauth/{provider}/start": {
      "get": {
        "summary": "开始 GitHub 或微信 OAuth 登录",
        "operationId": "startOAuthLogin",
        "parameters": [
          { "name": "provider", "in": "path", "required": true, "schema": { "type": "string", "enum": ["github", "wechat"] } }
        ],
        "responses": {
          "302": { "description": "跳转到第三方授权页" },
          "503": { "description": "登录提供方尚未配置", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/oauth/{provider}/callback": {
      "get": {
        "summary": "处理第三方 OAuth 回调并创建会话",
        "operationId": "completeOAuthLogin",
        "parameters": [
          { "name": "provider", "in": "path", "required": true, "schema": { "type": "string", "enum": ["github", "wechat"] } },
          { "name": "code", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "state", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "302": { "description": "登录成功并返回站点首页" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "503": { "description": "登录提供方尚未配置", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/favorites": {
      "get": {
        "summary": "获取当前用户收藏的项目 ID",
        "operationId": "listFavorites",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "收藏项目 ID 列表", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FavoriteListResponse" } } } },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/projects/{slug}/favorite": {
      "parameters": [
        { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "post": {
        "summary": "收藏项目",
        "operationId": "createFavorite",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "204": { "description": "已收藏；重复调用同样成功" },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "summary": "取消收藏项目",
        "operationId": "deleteFavorite",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "204": { "description": "已取消收藏；重复调用同样成功" },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/follows": {
      "get": {
        "summary": "获取当前用户关注的项目 ID",
        "operationId": "listFollows",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "关注项目 ID 列表", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FollowListResponse" } } } },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/projects/{slug}/follow": {
      "parameters": [
        { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "post": {
        "summary": "关注项目",
        "description": "关注后项目（重新）发布时会收到 project.updated 站内通知。重复调用同样成功（幂等）。",
        "operationId": "createFollow",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "204": { "description": "已关注；重复调用同样成功" },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "summary": "取消关注项目",
        "operationId": "deleteFollow",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "204": { "description": "已取消关注；重复调用同样成功" },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/projects/{slug}/share": {
      "parameters": [
        { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "post": {
        "summary": "记录项目分享并给分享者加经验",
        "description": "前端分享按钮在复制链接的同时调用；每项目每人只加一次经验（幂等）。",
        "operationId": "shareProject",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "204": { "description": "分享已记录；重复调用同样成功" },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/projects/{slug}/view": {
      "parameters": [
        { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "post": {
        "summary": "记录一次项目浏览（beacon）",
        "description": "前端打开项目详情时调用，累加浏览量。公开，无需登录。",
        "operationId": "recordProjectView",
        "responses": {
          "204": { "description": "浏览已记录" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/users/{id}/profile": {
      "get": {
        "summary": "获取用户公开主页",
        "description": "公开、无需登录。返回用户昵称、等级、注册时间、其已发布项目摘要与浏览/下载汇总。绝不返回邮箱或任何 PII。",
        "operationId": "getUserProfile",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "responses": {
          "200": {
            "description": "用户公开资料",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PublicUserProfileResponse" } } }
          },
          "404": { "description": "用户不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/notifications": {
      "get": {
        "summary": "获取当前用户站内通知",
        "operationId": "listNotifications",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "按时间倒序的通知列表和未读数量，最多返回 50 条", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NotificationListResponse" } } } },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/notifications/events": {
      "get": {
        "summary": "订阅当前用户通知实时事件",
        "description": "Server-Sent Events 通知流。事件名为 notification，数据为 Notification 对象；通知类型包括 comment.replied、project.approved、project.rejected、project.updated。",
        "operationId": "streamNotificationEvents",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Server-Sent Events 通知事件流", "content": { "text/event-stream": { "schema": { "type": "string" } } } },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/notifications/read-all": {
      "post": {
        "summary": "将当前用户全部通知标记为已读",
        "operationId": "markAllNotificationsRead",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "204": { "description": "全部通知已标记为已读" },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/notifications/{notificationID}/read": {
      "post": {
        "summary": "将单条通知标记为已读",
        "operationId": "markNotificationRead",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "notificationID", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "responses": {
          "204": { "description": "通知已标记为已读，重复操作幂等" },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "通知不存在或不属于当前用户", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/files/presign-upload": {
      "post": {
        "summary": "获取 OSS 文件直传授权",
        "operationId": "presignFileUpload",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ObjectUploadRequest" } } }
        },
        "responses": {
          "200": { "description": "十分钟有效的 OSS PUT 授权", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ObjectUploadResponse" } } } },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "422": { "description": "文件类型或大小不符合要求", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/files/author-asset": {
      "get": {
        "summary": "预览当前作者上传的正文资源",
        "operationId": "previewAuthorAsset",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "key", "in": "query", "required": true, "description": "Base64URL 编码的 OSS object key", "schema": { "type": "string" } }],
        "responses": {
          "307": { "description": "重定向到短时 OSS URL" },
          "401": { "description": "未登录", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/files/frame-asset": {
      "get": {
        "summary": "公开预览自定义头像框图片",
        "operationId": "previewFrameAsset",
        "parameters": [{ "name": "key", "in": "query", "required": true, "description": "Base64URL 编码的 OSS object key（须为 uploads/ 下的图片对象）", "schema": { "type": "string" } }],
        "responses": {
          "307": { "description": "重定向到短时 OSS URL" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "open_resouce_session"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Open API 密钥，形如 ork_<hex>"
      }
    },
...
    "schemas": {
...
      "AuthResponse": {
        "type": "object",
        "required": ["data", "request_id"],
        "properties": {
          "data": { "$ref": "#/components/schemas/AuthUser" },
          "request_id": { "type": "string" }
        }
      },
...
      "AcceptedResponse": {
        "type": "object",
        "required": ["request_id"],
        "properties": {
          "request_id": { "type": "string" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email", "display_name", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "maxLength": 254 },
          "display_name": { "type": "string", "minLength": 2, "maxLength": 80 },
          "password": { "type": "string", "minLength": 8, "maxLength": 128 }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "request_id"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "request_id": {
            "type": "string"
          }
        }
      }
    }
  }
}